enh: move comp to folder
